--- cornflow/commands/__init__.py
@@ -1,5 +1,0 @@
-"""
-Initialization file for the commands module
-"""
-
-from .commands import *
